+++ addon/components/yeti-table/tfoot.js
@@ -1,5 +1,4 @@
 import { tagName } from '@ember-decorators/component';
-import { deprecate } from '@ember/application/deprecations';
 import Component from '@ember/component';
 
 /**
@@ -24,19 +23,6 @@ class TFoot extends Component {
   parent;
 
   columns;
-
-  init() {
-    super.init(...arguments);
-
-    deprecate('The yielded `foot` component has been deprecated. Please use `tfoot` instead.', !this.shouldDeprecate, {
-      id: 'ember-yeti-table:foot-component',
-      until: '2.0.0',
-      for: 'ember-yeti-table',
-      since: {
-        enable: '1.4.0'
-      }
-    });
-  }
 }
 
 export default TFoot;